WIP more removal of redundant code
