+++ app/routers/verifier.py
@@ -2,7 +2,7 @@ import logging
 import time
 import traceback
 from typing import List, Optional
-from agent_factory import *
+from dependencies import *
 from facade import (
     # create_controller,
     get_schema_attributes,
