+++ app/routers/wallet.py
@@ -7,8 +7,7 @@ from aries_cloudcontroller import AriesAgentControllerBase
 from fastapi import APIRouter, Header, HTTPException, Depends
 
 from core.wallet import create_pub_did
-# from facade import create_controller
-from agent_factory import yoma_agent, member_agent
+from dependencies import yoma_agent, member_agent
 
 from schemas import (
     DidCreationResponse,
