+++ app/tests/conftest.py
@@ -5,11 +5,11 @@ from aries_cloudcontroller import AriesAgentControllerBase
 from aries_cloudcontroller.controllers.ledger import LedgerController
 from aries_cloudcontroller.controllers.wallet import WalletController
 from mockito import mock
-
-import agent_factory
+from fastapi import Depends
+# import agent_factory
 import ledger_facade
 import utils
-
+from dependencies import yoma_agent as agent
 
 @pytest.fixture
 def setup_env():
@@ -35,7 +35,7 @@ async def yoma_agent():
     # it is a bit of a pity that pytest fixtures don't do the same - I guess they want to maintain
     # flexibility - thus we have to.
     # this is doing what using decorators does for you
-    async with asynccontextmanager(agent_factory.yoma_agent)(
+    async with asynccontextmanager(agent)(
         x_api_key="adminApiKey"
     ) as c:
         yield c
