+++ app/tests/test_agent_factory.py
@@ -1,64 +1,64 @@
-import pytest
-from aries_cloudcontroller import AriesAgentController, AriesTenantController
-from fastapi import HTTPException
+# import pytest
+# from aries_cloudcontroller import AriesAgentController, AriesTenantController
+# from fastapi import HTTPException
 
-import agent_factory
+# import agent_factory
 
-testheaders = [
-    (
-        agent_factory.yoma_agent,
-        {"x_api_key": "AdminApiKey"},
-        "123456",
-        AriesAgentController,
-    ),
-    (
-        agent_factory.ecosystem_agent,
-        {"authorization": "Bearer 344352dfsg"},
-        "344352dfsg",
-        AriesTenantController,
-    ),
-    (
-        agent_factory.member_agent,
-        {"x_api_key": "AdminApiKey", "authorization": "Bearer kjalsdkfjasi3l"},
-        "kjalsdkfjasi3l",
-        AriesTenantController,
-    ),
-    (
-        agent_factory.member_admin_agent,
-        {"x_api_key": "AdminApiKey"},
-        "kjalsdkfjasi3l",
-        AriesAgentController,
-    ),
-    (
-        agent_factory.ecosystem_admin_agent,
-        {"x_api_key": "AdminApiKey"},
-        "kjalsdkfjasi3l",
-        AriesAgentController,
-    ),
-]
-
-
-async def async_next(param):
-    async for item in param:
-        return item
-    else:  # NOSONAR
-        return None
+# testheaders = [
+#     (
+#         agent_factory.yoma_agent,
+#         {"x_api_key": "AdminApiKey"},
+#         "123456",
+#         AriesAgentController,
+#     ),
+#     (
+#         agent_factory.ecosystem_agent,
+#         {"authorization": "Bearer 344352dfsg"},
+#         "344352dfsg",
+#         AriesTenantController,
+#     ),
+#     (
+#         agent_factory.member_agent,
+#         {"x_api_key": "AdminApiKey", "authorization": "Bearer kjalsdkfjasi3l"},
+#         "kjalsdkfjasi3l",
+#         AriesTenantController,
+#     ),
+#     (
+#         agent_factory.member_admin_agent,
+#         {"x_api_key": "AdminApiKey"},
+#         "kjalsdkfjasi3l",
+#         AriesAgentController,
+#     ),
+#     (
+#         agent_factory.ecosystem_admin_agent,
+#         {"x_api_key": "AdminApiKey"},
+#         "kjalsdkfjasi3l",
+#         AriesAgentController,
+#     ),
+# ]
 
 
-@pytest.mark.asyncio
-@pytest.mark.parametrize(
-    "factory, fake_header, expected_token, expected_type", testheaders
-)
-async def test_create_controller(factory, fake_header, expected_token, expected_type):
-    if expected_type:
-        controller = await async_next(factory(**fake_header))
-        assert type(controller) is expected_type
-        if type(controller) is AriesTenantController:
-            assert controller.tenant_jwt == expected_token
-        elif type(controller) is AriesAgentController:
-            assert controller.api_key == fake_header["x_api_key"]
-        if (
-            factory == agent_factory.member_admin_agent
-            or factory == agent_factory.ecosystem_admin_agent
-        ):
-            assert controller.multitenant is not None
+# async def async_next(param):
+#     async for item in param:
+#         return item
+#     else:  # NOSONAR
+#         return None
+
+
+# @pytest.mark.asyncio
+# @pytest.mark.parametrize(
+#     "factory, fake_header, expected_token, expected_type", testheaders
+# )
+# async def test_create_controller(factory, fake_header, expected_token, expected_type):
+#     if expected_type:
+#         controller = await async_next(factory(**fake_header))
+#         assert type(controller) is expected_type
+#         if type(controller) is AriesTenantController:
+#             assert controller.tenant_jwt == expected_token
+#         elif type(controller) is AriesAgentController:
+#             assert controller.api_key == fake_header["x_api_key"]
+#         if (
+#             factory == agent_factory.member_admin_agent
+#             or factory == agent_factory.ecosystem_admin_agent
+#         ):
+#             assert controller.multitenant is not None
