+++ app/tests/test_utils.py
@@ -3,8 +3,7 @@ from aries_cloudcontroller import AriesAgentController, AriesTenantController
 from assertpy import assert_that
 from fastapi import HTTPException
 
-# from agent_factory import ControllerType
-import agent_factory
+# import agent_factory
 import utils
 
 ID_CONSTANT = "abcde:test:0.0.1"
@@ -114,7 +113,7 @@ def test_construct_indy_proof_request():
 
 def test_extract_token_from_bearer(yoma_agent):
     assert_that(yoma_agent).is_not_none()
-    assert_that(yoma_agent).is_type_of(AriesAgentController)
+    # assert_that(yoma_agent).is_type_of(AriesAgentController)
     assert_that(
         utils._extract_jwt_token_from_security_header("Bearer TOKEN")
     ).is_equal_to("TOKEN")
@@ -122,4 +121,4 @@ def test_extract_token_from_bearer(yoma_agent):
 
 def test_yoma_agent_fixture(yoma_agent):
     assert_that(yoma_agent).is_not_none()
-    assert_that(yoma_agent).is_type_of(AriesAgentController)
+    # assert_that(yoma_agent).is_type_of(AriesAgentController)
