+++ app/utils.py
@@ -32,61 +32,6 @@ def _extract_jwt_token_from_security_header(jwt_token):
         raise HTTPException(401)
 
 
-# @deprecated(
-#     "logic moved to agent_factory - kept here because facade.create_controller still uses it"
-# )
-# def controller_factory(
-#     controller_type: ControllerType,
-#     x_api_key=None,
-#     authorization_header=None,
-#     x_wallet_id=None,
-# ) -> Type[Union[AriesAgentController, AriesTenantController]]:
-#     """
-#     Aries Controller factory returning an
-#     AriesController object based on a request header
-
-#     Parameters:
-#     -----------
-#     auth_headers: dict
-#         The header object containing wallet_id and jwt_token, or api_key
-
-#     Returns:
-#     --------
-#     controller: AriesCloudController (object)
-#     """
-#     if not controller_type:
-#         raise HTTPException(
-#             status_code=400,
-#             detail="Bad headers. Either provide an api_key or both wallet_id and tenant_jwt",
-#         )
-#     if controller_type == ControllerType.YOMA_AGENT:
-#         if not x_api_key:
-#             raise HTTPException(401)
-#         return AriesAgentController(
-#             admin_url=yoma_agent_url,
-#             api_key=x_api_key,
-#             is_multitenant=False,
-#         )
-#     elif controller_type == ControllerType.MEMBER_AGENT:
-#         if not authorization_header:
-#             raise HTTPException(401)
-#         return AriesTenantController(
-#             admin_url=member_agent_url,
-#             api_key=embedded_api_key,
-#             tenant_jwt=_extract_jwt_token_from_security_header(authorization_header),
-#             wallet_id=x_wallet_id,
-#         )
-#     elif controller_type == ControllerType.ECOSYSTEM_AGENT:
-#         if not authorization_header:
-#             raise HTTPException(401)
-#         return AriesTenantController(
-#             admin_url=ecosystem_agent_url,
-#             api_key=embedded_api_key,
-#             tenant_jwt=_extract_jwt_token_from_security_header(authorization_header),
-#             wallet_id=x_wallet_id,
-#         )
-
-
 def construct_zkp(zero_knowledge_proof: List[dict], schema_id: str) -> list:
     if zero_knowledge_proof == [{}]:
         return []
